requirements.txt
numpy~=1.23.3

PyAutoGUI~=0.9.53
Pillow~=9.2.0
